workaround for issue #628
Android API level 21 introduces class hierarchy change. The base class of
android.os.Bundle is changed from java.lang.Object to
android.os.BaseBundle and some of its methods are hoisted to the base
class. However, on some devices like Xiaomi Redmi BaseBundle is not super
class of Bundle as it is in android.jar from which we generate the
metadata. This causes JNI error when trying to call any of the hoisted
methods.