fix: build.gradle & kotlinVersion
Not sure if it is a change in Gradle Versions; but the "BeforeEvaluate"  function does NOT get called before the setting of the `kotlinVersion` in the top group.  So any external properties set in `gradle.properties` are NOT set at the point we check for it.   Copying the code to the top section in the `initialize` fixes the issue, however more tests will need to be done to see if we can just remove the beforeEvaluate section without any negative side effects.