chore: Switch from CreationContext() to GetCreationContextChecked()
See https://github.com/v8/v8/commit/b38bf5b0. The creation context may be
null for WASM objects, so the old API is deprecated in favour of
GetCreationContext() which returns a MaybeLocal<Context>. However, we are
not handling any WASM objects here, so we can use the version of the API
that asserts that we do have a creation context.