chore: Remove File::Exists()
This was primarily used by the heap snapshot code which is now removed.

It was also used internally in File::ReadBinary(), but that wasn't
necessary, as fopen() also handles the file not existing.

In new code, it's probably a good idea to perform operations without
checking if the file exists first anyway, as otherwise it can lead to race
conditions if the file is deleted.