chore: Move NativeScript-specific inspector files out of V8 sources folder
It's clearer that these files are part of NativeScript if they live in the
same folder as the other NativeScript sources. Otherwise, it's easy to
overlook them when updating the V8 sources and unintentionally delete
them.

The classes are adjusted slightly: moved from the v8_inspector:: namespace
into the tns:: namespace, and the filename casing updated to match the
other NativeScript sources.