fix: Implement console.log inspector with Runtime protocol
Console messages used to be implemented with the Log protocol, which is no
longer part of V8. The Runtime protocol provides an equivalent API,
Runtime.consoleAPICalled.

The Runtime protocol is part of the generated public sources, but the V8
inspector's RuntimeAgent is not. So, for now, we make use of V8's private
API, casting V8Inspector* to V8InspectorImpl* and V8InspectorSession* to
V8InspectorSessionImpl* in order to access their methods (for which we
pulled in the corresponding V8 header files in the V8 sources update,
earlier on this branch.)