docs: Add README steps
One step for installing the jsparser tests dependencies. This also updates
the lockfile if you do it with a recent version of NPM; speculatively
including that in this commit in case it doesn't break anything else.

Another step for setting the appropriate environment variables. They're
already mentioned in the Android Studio instructions above, but you'll
need to set them separately in your shell if building from the command
line.