test: Don't try to load lib/rt.jar in SBG tests
This JAR file, which contains the Java runtime classes, doesn't exist as a
separate file ever since Java 9.x. Therefore this test can't work as it
previously did.

This removes the setup code that tries to read rt.jar, which always fails.
It adds assertions for the results of generateBinding(), since in the
current state of the test, without access to rt.jar, the generation fails
and generateBinding() returns null.