chore: Don't store Global<Context> in callback info
Callbacks (runOnMainThread, frame, and timer) can run in the context in
which they were created, which we can get via GetCreationContext(). So,
there is no need to store a v8::Global<v8::Context> in which to invoke
the callback.

See https://github.com/v8/v8/commit/b38bf5b0 for why it's OK to use
ToLocalChecked() - the creation context is only null if the objects
were created in WASM instead of JS.