fix: Don't access `super` property on implementation object
Accessing `super` here in CallbackHandlers::GetMethodOverrides() would
cause a new empty super object to be created, and try to copy the
implementation object's metadata node over to it. This would fail a DCHECK
because the metadata node is null.

Instead, since we are checking in the next line if the property value is
a function, and we know that the value of `super` isn't a function, we can
instead just skip `super`.