fix: Avoid setting the same property on an ObjectTemplate twice
Due to Kotlin extension methods with the same names as Java methods, we
may be setting the same property twice on a class's prototype template.
The V8 API does not allow this and will fail a dcheck.

Since there is no API to check if a property has already been added to
v8::ObjectTemplate, wrap the prototype template in a small class
PrototypeTemplateFiller which keeps track of what's already been added and
avoids adding it again.