chore: Upgrade Gradle wrappers to 7.6
The Gradle build in the root folder uses Gradle 7.4 while test-app uses
Gradle 7.5. Upgrade both to use 7.6.

I have not looked too much into the Gradle 8.x series but it appears
there are a lot of features used that are deprecated in 8.x, so for now
this just upgrades to the latest in the 7.x series.