Gradle build improvements (#1780)
* chore: Upgrade Gradle wrappers to 7.6

The Gradle build in the root folder uses Gradle 7.4 while test-app uses
Gradle 7.5. Upgrade both to use 7.6.

I have not looked too much into the Gradle 8.x series but it appears
there are a lot of features used that are deprecated in 8.x, so for now
this just upgrades to the latest in the 7.x series.

* chore: Upgrade Gradle plugin

Android Studio complains about this every time the IDE is started, so we
may as well do what it asks. However, we cannot update to 8.0.0 or later
because that requires Gradle 8.0, so unfortunately we still have to live
with the error message.

* chore: Suppress unused parameter error messages in Kotlin tests

Android Studio complains about these when performing a gradle sync.

* chore: Move package namespace definitions into gradle files

According to Android Studio's build log, defining the package namespace in
AndroidManifest.xml is deprecated, and it should be done in build.gradle.

* chore: Remove Jetifier

Android Studio claims that Jetifier slows down the build, and that it is
not used in this project. It may have been used previously to update
un-updated external dependencies to use AndroidX, and that may not be
necessary anymore.

* chore: Replace JavaExec.main property with JavaExec.mainClass

The main property is deprecated and should be replaced with mainClass
which takes a Property<String> instance (created behind the scenes by
using '=') instead of a string.

* chore: keep jetifier enabled for now

---------

Co-authored-by: Osei Fortune <fortune.osei@yahoo.com>