chore: Don't copy generator .jar files to build-tools/ (#1789)
Gradle complains that when multiple tasks have the same output path, it
cannot decide accurately whether they need to run or not. The
copyJarToBuildTools task of all three generators (android-dts-generator,
android-metadata-generator, and static-binding-generator) shared the same
output path, so they were always getting run.

Instead, we can leave the .jar files in the location where they are built,
and reference them from there in tasks that subsequently require them.

This fix was submitted separately to android-dts-generator, which is a git
submodule, in https://github.com/NativeScript/android-dts-generator/pull/77
This commit pulls in a newer version of android-dts-generator which
includes that fix.